requirements.txt
django==3.1.7
djangorestframework==3.11.0
django-extensions==2.2.5
django_rq==2.3.2
django-rest-auth==0.9.3
docker==5.0.0
gunicorn
gevent